Major project directory refactoring
Created regex, helpers, and endpoints JS files for a more organized project directory

Refactored Commits screen component to account for the changes mentioned above (removed helper functions)

Updated types TS file to account for the changes mentioned above

Removed redundant import in BranchSelector component
