+++ client/src/misc/endpoints.js
@@ -1,0 +1,8 @@
+/**
+ * Only used for development builds. The local server
+ * endpoint URI to fetch commit data from.
+ */
+ export const LOCAL_SERVER_BASE_URL = "http://localhost:5000";
+
+ export const COMMITS_ENDPOINT = "/commits";
+ export const BRANCHES_ENDPOINT = "/branches";
