+++ client/src/screens/Commits.tsx
@@ -1,5 +1,7 @@
 import { Button } from "@material-ui/core";
 import React, { useCallback } from "react";
+
+// component imports
 import BranchSelector from "../components/BranchSelector";
 import CommitVisualizer from "../components/CommitVisualizer";
 import LinkField from "../components/LinkField";
@@ -15,24 +17,13 @@ import {
   setDefaultBranch,
 } from "../redux/reducers/repoBranchSlice";
 
-import { BranchData, BranchList, CommitArray } from "../types";
+// type imports
+import { BranchData, BranchList, CommitArray, ParsedLink } from "../types";
+
+// local helper function imports
+import { fetchBranches, fetchCommitData, parseLink } from "../misc/helpers";
 
 import "../styles/Commits.css"
-
-/** Regex for extracting the GitHub username from the link input. */
-const USER_NAME_REGEX = /(?<=https:\/\/github\.com\/)(.*)(?=\/)/;
-
-/** Regex for extracting the GitHub repo name from the link input. */
-const REPO_NAME_REGEX = /[^\/]+$/;
-
-/**
- * Only used for development builds. The local server
- * endpoint URI to fetch commit data from.
- */
-const LOCAL_SERVER_BASE_URL = "http://localhost:5000";
-
-const COMMITS_ENDPOINT = "/commits";
-const BRANCHES_ENDPOINT = "/branches";
 
 /**
  * Screen component responsible for visualizing GitHub
@@ -142,153 +133,6 @@ const Commits = () => {
       <CommitVisualizer />
     </div>
   );
-};
-
-export type ParsedLink = {
-  userName: string;
-  repoName: string;
-};
-
-/**
- * Parses the given link for a GitHub username and repository name.
- * Will throw an error if it fails to find either one.
- *
- * @param link The GitHub repo link to validate and parse.
- * @returns An object containing both username and repo name.
- */
-const parseLink = (link: string): ParsedLink => {
-  console.log("parseLink()");
-
-  const userName = link.match(USER_NAME_REGEX);
-  const repoName = link.match(REPO_NAME_REGEX);
-
-  // throw an error if there is no match for either
-  // the user or repo name
-  if (!userName && !repoName) {
-    // no user and repo name
-    throw new Error(
-      "Could not parse the given link for a username and repository name"
-    );
-  } else if (!userName) {
-    // no username only
-    throw new Error("Could not parse the given link for a username");
-  } else if (!repoName) {
-    // no repo name only
-    throw new Error("Could not parse the given link for a repository name");
-  } // end if else
-
-  console.log("end of parseLink()");
-
-  return {
-    userName: userName[0],
-    repoName: repoName[0],
-  };
-};
-
-/**
- * Fetches commit data of the default branch based on the given
- * GitHub username and repo name.
- *
- * @param userName The owner of the target repository.
- * @param repoName The target repository's name.
- * @param branch Optional. The branch to fetch commits from.
- * @returns The fetched commit data.
- */
-export const fetchCommitData = async (
-  userName: string,
-  repoName: string,
-  branch?: string
-): Promise<CommitArray> => {
-  console.log("fetchCommitData()");
-
-  // holds the fetched data
-  let fetchedData = [];
-
-  // console.log("parameters: ");
-  // console.table({ userName, repoName });
-
-  let targetUrl = `${productionEnvCheck(
-    COMMITS_ENDPOINT
-  )}?owner=${userName}&repo=${repoName}`;
-
-  if (branch) {
-    targetUrl += `&branch=${branch}`;
-  }
-
-  console.log("targetUrl: ", targetUrl);
-
-  try {
-    // IMPORTANT: a maximum of 30 commits can be fetched per request
-    // be sure to account for that later
-    const fetchedRes = await fetch(targetUrl);
-    // console.log("fetchedRes: ");
-    // console.log(fetchedRes);
-    const data = await fetchedRes.json();
-    // console.log("data: ");
-    // console.log(data);
-
-    if (data.message === "Not Found") {
-      throw new Error("Invalid link.")
-    }
-
-    fetchedData = data;
-  } catch (error) {
-    throw new Error(error);
-  }
-
-  console.log("end of fetchCommitData()");
-
-  return fetchedData;
-};
-
-/**
- * Fetches the branch list and default branch of the repository
- * based on the given GitHub username and repo name.
- *
- * @param userName The owner of the target repository.
- * @param repoName The target repository's name.
- * @returns The fetched branch list.
- */
-const fetchBranches = async (
-  userName: string,
-  repoName: string
-): Promise<any> => {
-  console.log("fetchBranchList()");
-
-  // holds the fetched branch list and default branch
-  let fetchedData: BranchData = { defaultBranch: "", branchList: [] };
-
-  const targetUrl = `${productionEnvCheck(
-    BRANCHES_ENDPOINT
-  )}?owner=${userName}&repo=${repoName}`;
-
-  try {
-    const fetchedRes = await fetch(targetUrl);
-    fetchedData = await fetchedRes.json();
-    console.log("fetched branches:");
-    console.log(fetchedData);
-  } catch (error) {
-    throw new Error(error);
-  }
-
-  console.log("end of fetchBranchList()");
-
-  return fetchedData;
-};
-
-/**
- * Helper function for target URL generation when fetching data.
- * Prepends localhost to the given endpoint if node environment
- * is not in production.
- *
- * @param endpoint The local server endpoint to request to.
- * @returns The endpoint by itself if in production environment; otherwise,
- * localhost:5000 prepended to the endpoint.
- */
-const productionEnvCheck = (endpoint: string): string => {
-  return process.env.NODE_ENV === "production"
-    ? endpoint
-    : LOCAL_SERVER_BASE_URL + endpoint;
 };
 
 export default Commits;
